Add counters for monitoring behaviour of the LMK
It turns out we can only monitor the three SG clocks without either extra
hardware or a lot of work: the MGT clocks to be received through some form
of MGT IP instance (the GTHE3_COMMON entity cannot be instantiated alone).

Also, the WCK is a bit tricky to get at, requiring a special constraint to
allow fabric routing of this clock.  Still, it works!