Revisit CA output alignment and clocking
First of all, it turns out that we need to use an MMCM for CK clock
regeneration if we want clocks to be in phase.

Next, it turns out to be much simpler to use clock phasing to get the
correct output alignment for CA, and so all the CA ODELAY can be removed.

Finally, CKEn is easy to align and goes back to being one bit in the
control interface.