Fudge to work around sticky CA6 bit -- PLEASE REVERT WHEN POSSIBLE!!!
THIS COMMIT **MUST** BE REVERTED AT THE FIRST OPPORTUNITY!!!

Add behaviour and control flag to trigger CA bus inversion (CABI) in
response to CA6 instead of counting bits.  This is used to work around the
observation that on the DLS prototype board the CA6 signal appears to be
stuck high.