Address a couple of issues triggered by ghdl
ghdl fails to correctly infer the required type context when assigning
elements of a vector_array, and the workaround is straighforward.

Similarly ghdl doesn't like the use of selected signal assignment in a
process.  As this is probably technically invalid and the fix is clean
this is also fixed here.