Add process for building GDDR6 interface as IP
This is far more involved than it ought to be.  The goal here is to make it
easy to use the memory controller as part of the Vivado IP Integrator
(Block Design editor).  The interfaces are defined using X_INTERFACE_INFO
attributes in the top level file gddr6_ip.vhd and the IP is assembled by
running make in the gddr6 directory.

Unfortunately it turns out that Vivado IP doesn't support the use of VHDL
2008 constructs in *any* files in the IP.  I work around this sad state of
affairs by generating a netlist instead and integrating that into the
generated IP.  Unfortunately this rules out any configuration options when
instantiating the IP.