Add support for fine control over the CA output phase
This is part of an investigation into reproducibility problems with the CK
phase: it seems that increased loading on CK changes the internal phase
offset quite substantially ... the BUFGCE on the MMCM feedback loop isn't
correcting this.

Unfortunately implementing this means that the CK delay clock phase will
need to be manually set as part of startup training.