Working on IP timing
Clock alignment seems to be a bit of a can of worms, and there are some key
differences between the IP build and a more general component build.  This
will need further research.

For the moment a special constraint file applying a CLOCK_DELAY_GROUP is
applied to the IP build, but is not applied in the final build: at the
moment it seems to be enough to just put ck_clk_delay on the the feedback
aligned clock.  Actually applying this constraint will prevent read
training from succeeding.