Support 49 bit DMA width and expand SGRAM interface to full 32-bit address
The driver now supports prom configuration with RW access and with support
for 64 bit addressing.  At the same time we need to change the external
address width to 49 bits.

A lot of trouble with the IP integrator seems to be sidestepped by forcing
the width and mode settings for the AXI Data Width converter: relying on
inference seems to trigger a *lot* of IP integrator misbehaviour.