Add LMK status polling to lmk04616 support
This needs to be explicitly enabled by setting a .ENABLE_STATUS register
flag.

Also take proper account of LMK multiplexing switching time.  According to
the TI documentation for the TXBN0304 (used for access control to the LMK)
it can take up to 172ns for switching to complete, so we have to wait this
out.